Project Finished; Ready for Review
In record time, I have completed my first JavaScript project.

This included linking the file that would contain a series of functions that told the index what type of image it should choose appropriate to the user's current time (also programmed to appear with JS).

With the speed I was able to get through by adding comments as I went and getting help from a certain someone (*cough* Micah *cough*), this project was quite the breeze. Onto the next one!